refactor(go): add comm.Interval
